finalizado

+++ src/main/java/br/com/sabium/controller/CursoController.java
@@ -21,17 +21,24 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import br.com.sabium.controller.dto.CursoDTO;
+import br.com.sabium.controller.dto.DisciplinaDTO;
+import br.com.sabium.controller.exception.CursoWithDisciplinasAssociateException;
 import br.com.sabium.controller.form.CursoForm;
 import br.com.sabium.model.administrativo.Curso;
+import br.com.sabium.model.administrativo.Disciplina;
 import br.com.sabium.repository.CursoRepository;
+import br.com.sabium.repository.DisciplinaRepository;
 
 @RestController
 @RequestMapping("/cursos")
 public class CursoController {
-	
+
 	@Autowired
 	private CursoRepository cursoRepository;
-	
+
+	@Autowired
+	private DisciplinaRepository disciplinaRepository;
+
 	@GetMapping
 	public List<CursoDTO> lista(String nomeCurso) {
 		List<Curso> cursos = new ArrayList<>();
@@ -39,31 +46,37 @@ public class CursoController {
 			cursos = cursoRepository.findAll();
 		} else {
 			cursos.add(cursoRepository.findByNome(nomeCurso));
-		}			
+		}
 		return CursoDTO.converter(cursos);
 	}
-	
+
 	@PostMapping
 	@Transactional
-	public ResponseEntity<CursoDTO> cadastrar(@RequestBody @Valid CursoForm form,
-			UriComponentsBuilder uriBuilder) {
+	public ResponseEntity<CursoDTO> cadastrar(@RequestBody @Valid CursoForm form, UriComponentsBuilder uriBuilder) {
 		Curso curso = form.converter(cursoRepository);
 		cursoRepository.save(curso);
-		
+
 		URI uri = uriBuilder.path("/cursos/{id}").buildAndExpand(curso.getId()).toUri();
 		return ResponseEntity.created(uri).body(new CursoDTO(curso));
 	}
-	
-	
+
 	@GetMapping("/{id}")
 	public ResponseEntity<CursoDTO> detalhar(@PathVariable Long id) {
 		Optional<Curso> curso = cursoRepository.findById(id);
 		if (curso != null) {
-			return ResponseEntity.ok(new CursoDTO(curso.get()));
+			CursoDTO cursoDTO = new CursoDTO(curso.get());
+			List<Disciplina> disciplinasOptional = disciplinaRepository.findByCursoId(curso.get().getId());
+			if (!disciplinasOptional.isEmpty()) {
+				List<Disciplina> disciplinas = new ArrayList<Disciplina>();
+				disciplinasOptional.forEach(d -> disciplinas.add(d));
+				List<DisciplinaDTO> dtos = DisciplinaDTO.converter(disciplinas);
+				cursoDTO.setDisciplinas(dtos);
+			}
+			return ResponseEntity.ok(cursoDTO);
 		}
 		return ResponseEntity.notFound().build();
 	}
-	
+
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<CursoDTO> atualizar(@PathVariable Long id, @RequestBody @Valid CursoForm form) {
@@ -72,31 +85,22 @@ public class CursoController {
 			Curso curso = form.atualizar(id, cursoRepository);
 			return ResponseEntity.ok(new CursoDTO(curso));
 		}
-		
+
 		return ResponseEntity.notFound().build();
 	}
-	
+
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
 		Optional<Curso> optional = cursoRepository.findById(id);
 		if (optional.isPresent()) {
-			cursoRepository.deleteById(id);
-			return ResponseEntity.ok().build();
+			if (optional.get().getDisciplinas().isEmpty()) {
+				cursoRepository.deleteById(id);
+				return ResponseEntity.ok().build();
+			}
+			throw new CursoWithDisciplinasAssociateException();
 		}
 		return ResponseEntity.notFound().build();
 	}
 
-	
 }
-
-
-
-
-
-
-
-
-
-
-
